# backend/src/Api/Controllers/RecordsController.cs
using FamilyVeda.Application.Common;
using FamilyVeda.Application.Records;
using FamilyVeda.Domain.Common;
using FluentValidation;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.RateLimiting;

namespace FamilyVeda.Api.Controllers;

[Route("api/v1")]
[Authorize(Policy = "FamilyUser")]
public sealed class RecordsController(IRecordService recordService, ILabExtractionService extractionService) : ApiControllerBase
{
    [HttpGet("members/{memberId:guid}/records")]
    public async Task<ActionResult<PagedResult<HealthRecordDto>>> GetRecords(Guid memberId, int page = 1, int pageSize = 20, string? search = null, RecordType? type = null, string sort = "newest", CancellationToken cancellationToken = default) =>
        Ok(await recordService.GetRecordsAsync(memberId, page, pageSize, search, type, sort, cancellationToken));

    [HttpPost("members/{memberId:guid}/records")]
    public async Task<ActionResult<HealthRecordDto>> CreateRecord(Guid memberId, UpsertHealthRecordRequest request, IValidator<UpsertHealthRecordRequest> validator, CancellationToken cancellationToken)
    {
        await ValidateAsync(validator, request, cancellationToken);
        var record = await recordService.CreateRecordAsync(memberId, request, cancellationToken);
        return Created($"/api/v1/records/{record.Id}", record);
    }

    [HttpPut("records/{recordId:guid}")]
    public async Task<ActionResult<HealthRecordDto>> UpdateRecord(Guid recordId, UpsertHealthRecordRequest request, IValidator<UpsertHealthRecordRequest> validator, CancellationToken cancellationToken)
    {
        await ValidateAsync(validator, request, cancellationToken);
        return Ok(await recordService.UpdateRecordAsync(recordId, request, cancellationToken));
    }

    [HttpDelete("records/{recordId:guid}")]
    public async Task<ActionResult<object>> DeleteRecord(Guid recordId, CancellationToken cancellationToken)
    {
        await recordService.DeleteRecordAsync(recordId, cancellationToken);
        return NoContent();
    }

    [HttpGet("members/{memberId:guid}/vitals")]
    public async Task<ActionResult<IReadOnlyList<VitalDto>>> GetVitals(Guid memberId, CancellationToken cancellationToken) =>
        Ok(await recordService.GetVitalsAsync(memberId, cancellationToken));

    [HttpPost("members/{memberId:guid}/vitals")]
    public async Task<ActionResult<VitalDto>> AddVital(Guid memberId, CreateVitalRequest request, IValidator<CreateVitalRequest> validator, CancellationToken cancellationToken)
    {
        await ValidateAsync(validator, request, cancellationToken);
        var vital = await recordService.AddVitalAsync(memberId, request, cancellationToken);
        return Created($"/api/v1/members/{memberId}/vitals/{vital.Id}", vital);
    }

    [HttpGet("members/{memberId:guid}/vitals/trends")]
    public async Task<ActionResult<IReadOnlyList<VitalTrendDto>>> GetVitalTrends(Guid memberId, CancellationToken cancellationToken) =>
        Ok(await recordService.GetVitalTrendsAsync(memberId, cancellationToken));

    [HttpGet("members/{memberId:guid}/lab-reports")]
    public async Task<ActionResult<IReadOnlyList<LabReportDto>>> GetLabReports(Guid memberId, CancellationToken cancellationToken) =>
        Ok(await recordService.GetLabReportsAsync(memberId, cancellationToken));

    [HttpGet("lab-reports/{reportId:guid}")]
    public async Task<ActionResult<LabReportDetailDto>> GetLabReport(Guid reportId, CancellationToken cancellationToken) =>
        Ok(await recordService.GetLabReportDetailAsync(reportId, cancellationToken));

    [HttpPut("lab-reports/{reportId:guid}/review")]
    public async Task<ActionResult<LabReportDetailDto>> ConfirmLabReport(Guid reportId, ConfirmLabReportRequest request, CancellationToken cancellationToken) =>
        Ok(await recordService.ConfirmLabReportAsync(reportId, request, cancellationToken));

    [HttpPost("members/{memberId:guid}/lab-reports")]
    [Consumes("multipart/form-data")]
    [RequestFormLimits(MultipartBodyLengthLimit = 10_485_760)]
    public async Task<ActionResult<LabReportDto>> UploadLabReport(Guid memberId, [FromForm] LabReportUploadForm form, CancellationToken cancellationToken)
    {
        await using var stream = form.File.OpenReadStream();
        var report = await recordService.UploadLabReportAsync(memberId, form.File.FileName, form.File.ContentType, form.File.Length, stream, form.CollectedAt, cancellationToken);
        return Created($"/api/v1/lab-reports/{report.Id}", report);
    }

    [HttpPost("lab-reports/{reportId:guid}/extract")]
    [EnableRateLimiting("Ocr")]
    public async Task<ActionResult<LabExtractionResultDto>> ExtractLabReport(Guid reportId, CancellationToken cancellationToken) =>
        Ok(await extractionService.ExtractAsync(reportId, cancellationToken));

    [HttpGet("members/{memberId:guid}/hereditary-flags")]
    public async Task<ActionResult<IReadOnlyList<HereditaryFlagDto>>> GetHereditaryFlags(Guid memberId, CancellationToken cancellationToken) =>
        Ok(await recordService.GetHereditaryFlagsAsync(memberId, cancellationToken));
}

public sealed class LabReportUploadForm
{
    public required IFormFile File { get; init; }
    public DateTimeOffset? CollectedAt { get; init; }
}