# docs/api/openapi.yaml
...
openapi: 3.0.1
info:
  title: Family Veda API
  version: v1
paths:
...
  "/api/v1/members/{memberId}/records":
    get:
      tags:
      - Records
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: page
        in: query
        schema:
          type: integer
          format: int32
          default: 1
      - name: pageSize
        in: query
        schema:
          type: integer
          format: int32
          default: 20
      - name: search
        in: query
        schema:
          type: string
      - name: type
        in: query
        schema:
          "$ref": "#/components/schemas/RecordType"
      - name: sort
        in: query
        schema:
          type: string
          default: newest
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HealthRecordDtoPagedResult"
    post:
      tags:
      - Records
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpsertHealthRecordRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/UpsertHealthRecordRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/UpsertHealthRecordRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HealthRecordDto"
  "/api/v1/records/{recordId}":
    put:
      tags:
      - Records
      parameters:
      - name: recordId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/UpsertHealthRecordRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/UpsertHealthRecordRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/UpsertHealthRecordRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HealthRecordDto"
    delete:
      tags:
      - Records
      parameters:
      - name: recordId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
  "/api/v1/members/{memberId}/vitals":
    get:
      tags:
      - Records
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/VitalDto"
    post:
      tags:
      - Records
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CreateVitalRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/CreateVitalRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/CreateVitalRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/VitalDto"
  "/api/v1/members/{memberId}/vitals/trends":
    get:
      tags:
      - Records
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/VitalTrendDto"
  "/api/v1/members/{memberId}/lab-reports":
    get:
      tags:
      - Records
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/LabReportDto"
    post:
      tags:
      - Records
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
              - File
              type: object
              properties:
                File:
                  type: string
                  format: binary
                CollectedAt:
                  type: string
                  format: date-time
            encoding:
              File:
                style: form
              CollectedAt:
                style: form
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/LabReportDto"
  "/api/v1/lab-reports/{reportId}":
    get:
      tags:
      - Records
      parameters:
      - name: reportId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/LabReportDetailDto"
  "/api/v1/lab-reports/{reportId}/review":
    put:
      tags:
      - Records
      parameters:
      - name: reportId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ConfirmLabReportRequest"
          text/json:
            schema:
              "$ref": "#/components/schemas/ConfirmLabReportRequest"
          application/*+json:
            schema:
              "$ref": "#/components/schemas/ConfirmLabReportRequest"
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/LabReportDetailDto"
  "/api/v1/lab-reports/{reportId}/extract":
    post:
      tags:
      - Records
      parameters:
      - name: reportId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/LabExtractionResultDto"
  "/api/v1/members/{memberId}/hereditary-flags":
    get:
      tags:
      - Records
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/HereditaryFlagDto"
...
components:
  schemas:
...
    ConfirmLabReportRequest:
      type: object
      properties:
        values:
          type: array
          items:
            "$ref": "#/components/schemas/ConfirmLabValueRequest"
          nullable: true
        confirmedFlagIds:
          type: array
          items:
            type: string
            format: uuid
          nullable: true
      additionalProperties: false
    ConfirmLabValueRequest:
      type: object
      properties:
        id:
          type: string
          format: uuid
        analyte:
          type: string
          nullable: true
        value:
          type: number
          format: double
        unit:
          type: string
          nullable: true
        referenceLow:
          type: number
          format: double
          nullable: true
        referenceHigh:
          type: number
          format: double
          nullable: true
      additionalProperties: false
...
    CreateVitalRequest:
      type: object
      properties:
        vitalType:
          type: string
          nullable: true
        value:
          type: number
          format: double
        unit:
          type: string
          nullable: true
        measuredAt:
          type: string
          format: date-time
      additionalProperties: false
...
    HealthRecordDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        memberId:
          type: string
          format: uuid
        recordType:
          "$ref": "#/components/schemas/RecordType"
        title:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
        occurredOn:
          type: string
          format: date
      additionalProperties: false
    HealthRecordDtoPagedResult:
      type: object
      properties:
        items:
          type: array
          items:
            "$ref": "#/components/schemas/HealthRecordDto"
          nullable: true
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        totalCount:
          type: integer
          format: int32
        totalPages:
          type: integer
          format: int32
          readOnly: true
      additionalProperties: false
    HereditaryFlagDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        memberId:
          type: string
          format: uuid
        conditionCode:
          type: string
          nullable: true
        finding:
          type: string
          nullable: true
        confidence:
          type: number
          format: double
        manuallyConfirmed:
          type: boolean
      additionalProperties: false
    LabExtractionResultDto:
      type: object
      properties:
        reportId:
          type: string
          format: uuid
        status:
          "$ref": "#/components/schemas/OcrStatus"
        valuesExtracted:
          type: integer
          format: int32
        flagsExtracted:
          type: integer
          format: int32
        requiresManualReview:
          type: boolean
      additionalProperties: false
    LabReportDetailDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        memberId:
          type: string
          format: uuid
        originalFileName:
          type: string
          nullable: true
        ocrStatus:
          "$ref": "#/components/schemas/OcrStatus"
        collectedAt:
          type: string
          format: date-time
          nullable: true
        values:
          type: array
          items:
            "$ref": "#/components/schemas/LabValueDto"
          nullable: true
        flags:
          type: array
          items:
            "$ref": "#/components/schemas/HereditaryFlagDto"
          nullable: true
      additionalProperties: false
    LabReportDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        memberId:
          type: string
          format: uuid
        originalFileName:
          type: string
          nullable: true
        ocrStatus:
          "$ref": "#/components/schemas/OcrStatus"
        collectedAt:
          type: string
          format: date-time
          nullable: true
      additionalProperties: false
    LabValueDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        analyte:
          type: string
          nullable: true
        value:
          type: number
          format: double
        unit:
          type: string
          nullable: true
        referenceLow:
          type: number
          format: double
          nullable: true
        referenceHigh:
          type: number
          format: double
          nullable: true
        wasManuallyConfirmed:
          type: boolean
      additionalProperties: false
...
    OcrStatus:
      enum:
      - Pending
      - Processing
      - Completed
      - Failed
      - ManualEntry
      type: string
    RecordType:
      enum:
      - Condition
      - Allergy
      - Medication
      - Surgery
      - Note
      type: string
...
    UpsertHealthRecordRequest:
      type: object
      properties:
        recordType:
          "$ref": "#/components/schemas/RecordType"
        title:
          type: string
          nullable: true
        summary:
          type: string
          nullable: true
        occurredOn:
          type: string
          format: date
      additionalProperties: false
...
    VitalDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        memberId:
          type: string
          format: uuid
        vitalType:
          type: string
          nullable: true
        value:
          type: number
          format: double
        unit:
          type: string
          nullable: true
        measuredAt:
          type: string
          format: date-time
      additionalProperties: false
    VitalPointDto:
      type: object
      properties:
        measuredAt:
          type: string
          format: date-time
        value:
          type: number
          format: double
        unit:
          type: string
          nullable: true
      additionalProperties: false
    VitalTrendDto:
      type: object
      properties:
        vitalType:
          type: string
          nullable: true
        points:
          type: array
          items:
            "$ref": "#/components/schemas/VitalPointDto"
          nullable: true
      additionalProperties: false
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
- Bearer: []